Start replacing sharedpref usages with settings wrapper
