+++ app/build.gradle
@@ -69,6 +69,7 @@ android {
 dependencies {
     implementation project(':ui')
     implementation project(':strings')
+    implementation project(':settings')
     implementation project(':sharing')
     implementation project(':testing')
