+++ app/src/main/java/journal/gratitude/com/gratitudejournal/di/ApplicationModule.kt
@@ -1,17 +1,13 @@
 package journal.gratitude.com.gratitudejournal.di
 
 import android.content.Context
-import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import journal.gratitude.com.gratitudejournal.repository.EntryRepository
-import journal.gratitude.com.gratitudejournal.repository.EntryRepositoryImpl
 import journal.gratitude.com.gratitudejournal.room.EntryDao
 import journal.gratitude.com.gratitudejournal.room.EntryDatabase
-import javax.inject.Singleton
 
 
 @Module
