+++ app/src/main/java/journal/gratitude/com/gratitudejournal/di/SettingsEntryPoint.kt
@@ -1,0 +1,12 @@
+package journal.gratitude.com.gratitudejournal.di
+
+import com.presently.settings.PresentlySettings
+import dagger.hilt.EntryPoint
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@EntryPoint
+@InstallIn(SingletonComponent::class)
+interface SettingsEntryPoint {
+    val settings: PresentlySettings
+}
